:hammer:Update time to long

+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dao/JobInstanceDAO.java
@@ -31,5 +31,5 @@ public interface JobInstanceDAO {
      * @param lastReportTime lastReportTime
      * @return Integer
      */
-    Integer updateLastReportTimeById(Long id, Integer lastReportTime);
+    Integer updateLastReportTimeById(Long id, Long lastReportTime);
 }
